MAINT: stray print/lint
* While investigating gh-48, I noticed a stray debug `print()` call
accidentally slipped in to `test_regression_boston()`. This patch
proposes to turn on an additional `ruff` rule to automatically
detect these cases in the future, fixes the one offense in our
codebase, and adds an exception for the usage in `scripts` (which
we don't hold to the same standard at the moment).