WIP, DOC: start adding numpydoc docstrings
* Related to the addition of docstrings mentioned in gh-45.

* This is very much a work in progress for now. It proposes
to conform to the numpydoc docstring standard, add a linter
for this standard in the CI, and add the first few docstrings.
Until more functions/classes are brought up to standard, we expect
the CI/doc linting to fail here.

* Many more docstrings need to be added of course--feel free
to push in some drafts, I'll chip away at them when I can.

* This does not yet tackle the task of turning these docstrings
into HTML files the public can access, but it is a precursor
to that.